fixed jwt role issue

+++ Controllers/CartsController.cs
@@ -6,10 +6,12 @@ using System.Data.Entity.Infrastructure;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using System.Web.Http;
 using System.Web.Http.Description;
 using FurniflexBE.Context;
+using FurniflexBE.Helpers;
 using FurniflexBE.Models;
 
 namespace FurniflexBE.Controllers
@@ -20,9 +22,22 @@ namespace FurniflexBE.Controllers
         private AppDbContext db = new AppDbContext();
 
         // GET: api/Carts
-        public IQueryable<Cart> Getcarts()
+        [HttpGet]
+        [ResponseType(typeof(List<Cart>))]
+        public IHttpActionResult Getcarts()
         {
-            return db.carts;
+
+            var userRole = IdentityHelper.GetRoleName(User.Identity as ClaimsIdentity);
+
+            if (userRole != "admin")
+            {
+                return Unauthorized(); // Returns 401 Unauthorized if the user is not an admin
+            }
+
+            var carts = db.carts.ToList(); // Get all orders as a list
+
+            return Ok(carts);
+       
         }
 
         // GET: api/Carts/5
@@ -42,13 +57,26 @@ namespace FurniflexBE.Controllers
         [ResponseType(typeof(void))]
         public async Task<IHttpActionResult> PutCart(int id, int changeValue)
         {
-
+            var identityId = IdentityHelper.GetUserId(User.Identity as ClaimsIdentity);
+            
+            if (identityId == null)
+            {
+                return Unauthorized();
+            }
+            
 
             var cart =await db.carts.FindAsync(id);
             if(cart == null)
             {
                 return NotFound();
             }
+
+            var userId = cart.UserId;
+            if (userId != identityId)
+            {
+                return BadRequest("Cart is not Yours, You cannot modify");
+            }
+
 
             cart.Quantity += changeValue;
 
@@ -144,6 +172,16 @@ namespace FurniflexBE.Controllers
         [Route("api/Carts/DeleteByUserId/{userId}")]
         public async Task<IHttpActionResult> DeleteCartByUserId(int userId)
         {
+            var identityId = IdentityHelper.GetUserId(User.Identity as ClaimsIdentity);
+            var userRole = IdentityHelper.GetRoleName(User.Identity as ClaimsIdentity);
+            if (identityId == null)
+            {
+                return Unauthorized();
+            }
+            if (userId != identityId && userRole != "admin")
+            {
+                return BadRequest("Cart is not Yours, nor you are an admin. You can't delete the cart");
+            }
             var carts = db.carts.Where(c => c.UserId == userId).ToList();
 
             if (!carts.Any())
@@ -161,13 +199,24 @@ namespace FurniflexBE.Controllers
         [ResponseType(typeof(Cart))]
         public async Task<IHttpActionResult> DeleteCart(int id)
         {
+            var userId = IdentityHelper.GetUserId(User.Identity as ClaimsIdentity);
+            var userRole = IdentityHelper.GetRoleName(User.Identity as ClaimsIdentity);
+            if (userId == null)
+            {
+                return Unauthorized();
+            }
+
             Cart cart = await db.carts.FindAsync(id);
             if (cart == null)
             {
                 return NotFound();
             }
 
-            var userId = cart.User.UserId;
+            var cartUserId = cart.User.UserId;
+            if (userId != cartUserId && userRole !="admin")
+            {
+                return BadRequest("Cart is not Yours, nor you are an admin. You can't delete the cart");
+            }
 
             db.carts.Remove(cart);
             await db.SaveChangesAsync();
